add data download routine

+++ Source/Dataloader/MixData.py
@@ -110,13 +110,13 @@ class MixData(Dataset):
         if self.era5_network is not None:
             era5_u = torch.from_numpy(
                 np.moveaxis(self.era5_data.u10.sel(time=slice(time_sel[0], time_sel[-1])).values, 0, -1).reshape(
-                    (self.era5_network.era5_pos.size(0), -1)))
+                    (self.era5_network.era5_pos.size(0), -1)).astype(np.float32))
             era5_v = torch.from_numpy(
                 np.moveaxis(self.era5_data.v10.sel(time=slice(time_sel[0], time_sel[-1])).values, 0, -1).reshape(
-                    (self.era5_network.era5_pos.size(0), -1)))
+                    (self.era5_network.era5_pos.size(0), -1)).astype(np.float32))
             era5_temp = torch.from_numpy(
                 np.moveaxis(self.era5_data.t2m.sel(time=slice(time_sel[0], time_sel[-1])).values, 0, -1).reshape(
-                    (self.era5_network.era5_pos.size(0), -1)))
+                    (self.era5_network.era5_pos.size(0), -1)).astype(np.float32))
 
             sample[f'e2m_edge_index'] = self.era5_network.e2m_edge_index
             sample[f'era5_u'] = era5_u
